GarageControl/requirements.txt
django>=5.2,<5.3
requests>=2.28.0
bleak>=0.20.0
python-dotenv>=1.0.0
pandas>=2.1.0
matplotlib>=3.7.1
numpy>=1.24.0
# Weather Underground scraping dependencies
selenium>=4.15.0
webdriver-manager>=4.0.0 
bleak==0.21.1